requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==2.2.2
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-openai==0.3.24
langchain-community==0.3.26
rank_bm25==0.2.2
-e .
